+++ src/main/java/com/sparta/hanghae_homework_week04/service/NoticeBoardService.java
@@ -1,10 +1,12 @@
 package com.sparta.hanghae_homework_week04.service;
 
 import com.sparta.hanghae_homework_week04.domain.NoticeBoard;
-import com.sparta.hanghae_homework_week04.dto.NoticeBoardDto;
+import com.sparta.hanghae_homework_week04.domain.User;
+import com.sparta.hanghae_homework_week04.dto.NoticeBoardRequestDto;
+import com.sparta.hanghae_homework_week04.repository.CommentRepository;
 import com.sparta.hanghae_homework_week04.repository.NoticeBoardRepository;
 import lombok.RequiredArgsConstructor;
-import org.aspectj.weaver.ast.Not;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 
@@ -18,41 +20,49 @@ public class NoticeBoardService {
 
     private final NoticeBoardRepository noticeBoardRepository;
 
-    public List<NoticeBoardDto> findAll() {
+    private final CommentRepository commentRepository;
+
+    public List<NoticeBoardRequestDto> findAll() {
 
         List<NoticeBoard> noticeBoards = noticeBoardRepository.findAll();
-        List<NoticeBoardDto> noticeBoardDtos = new ArrayList<>();
+        List<NoticeBoardRequestDto> noticeBoardRequestDtos = new ArrayList<>();
 
         for (NoticeBoard noticeBoard : noticeBoards) {
 
-            NoticeBoardDto boardDto = new NoticeBoardDto(noticeBoard);
+            NoticeBoardRequestDto boardDto = new NoticeBoardRequestDto(noticeBoard);
 
-            noticeBoardDtos.add(boardDto);
+            noticeBoardRequestDtos.add(boardDto);
         }
 
-        return noticeBoardDtos;
+        return noticeBoardRequestDtos;
     }
 
-    public NoticeBoardDto getPost(Long id) {
+    public NoticeBoardRequestDto getPost(Long id) {
 
         NoticeBoard noticeBoard = noticeBoardRepository.findById(id).orElseThrow(
                 () -> new RuntimeException("Board ID does not exist"));
 
-        NoticeBoardDto noticeBoardDto = new NoticeBoardDto(noticeBoard);
+        NoticeBoardRequestDto noticeBoardRequestDto = new NoticeBoardRequestDto(noticeBoard);
 
-        return noticeBoardDto;
+        return noticeBoardRequestDto;
 
     }
 
-    public void save(NoticeBoardDto noticeBoardDto) {
+    public void save(NoticeBoardRequestDto noticeBoardRequestDto, User user) {
 
-        NoticeBoard board = new NoticeBoard(noticeBoardDto);
+        NoticeBoard board = new NoticeBoard(noticeBoardRequestDto, user);
 
         noticeBoardRepository.save(board);
     }
 
     @Transactional
-    public Long update(NoticeBoardDto requestDto, long requestId) {
+    public Long update(NoticeBoardRequestDto requestDto, long requestId, User user) {
+        
+        User updateUser = noticeBoardRepository.findById(requestId).get().getUser();
+        
+        if(!user.equals(updateUser)){
+            throw new RuntimeException("게시글 작성자와 Login User가 일치하지 않습니다.");
+        }
 
         NoticeBoard noticeBoard = noticeBoardRepository.findById(requestId).orElseThrow(
                 () -> new RuntimeException("Post ID does not exist")
@@ -65,9 +75,17 @@ public class NoticeBoardService {
     }
 
     @Transactional
-    public Long deletePost(Long id) {
+    public Long deletePost(Long id, User user) {
+
+        User deleteUser = noticeBoardRepository.findById(id).get().getUser();
+
+        if(!user.equals(deleteUser)){
+            throw new RuntimeException("게시글 작성자와 Login User가 일치하지 않습니다.");
+        }
 
         noticeBoardRepository.deleteById(id);
+        commentRepository.deleteByNoticeBoard(id);
+
 
         return id;
     }
